academic-copilot/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
google-genai==1.14.0
google-auth==2.40.1
google-auth-oauthlib==1.2.1
google-api-python-client==2.166.0
httpx==0.28.1
python-dotenv==1.1.0
python-multipart==0.0.20
